Change in var storage
I’m switching to a dictionary to store all of my variables. this
dictionary is called ‘Settings’ and will include all parameters except
raw data or output files. this dictionary will be saved as out as a CSV
with a unique time stamp, so tat each iteration of the analysis has a
record.

#16683064947992