Settings save
now can save settings out. used a pandas df to hold and organize it,
since writing it out was a lot easier than doing it right out of the
dictionary. the date time stamp is in the file save name AND the file
itself, so that this file can never be saved over.

#16683064947992