event psd and pyeeg fix
event psd was taking the inverse of a number that was already in hertz.
however, the ‘function of sample rate’ plan for psd doesn’t make a lot
of sense in the graphs AND the power in bands don’t even calculate.
something must be broken now that we aren’t using 256 and 128 as
defaults.

pyeeg wasn’t being imported properly in the functions ap_ent wrapper
and samp_ent wrapper. now fixed. also updated blocks on raw signal
entropy.

moved documentation inside the python file.
#32388207689197